refactor(validator): move validator package to participants
